requirements.txt
# Core HTTP and async
httpx==0.27.2

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-html==4.1.1

# Validation and configuration
pydantic==1.10.17
python-dotenv==1.0.1

# Retry and logging
tenacity==9.0.0
rich==13.9.2
